test: move tests to src directory
