+++ packages/standard-validator/src/__schemas__/arktype.ts
