+++ packages/swagger-ui/src/option-renderer.test.ts
@@ -1,7 +1,7 @@
 /*eslint quotes: ["off", "single"]*/
 
-import type { DistSwaggerUIOptions } from '../src/swagger/renderer'
-import { renderSwaggerUIOptions } from '../src/swagger/renderer'
+import type { DistSwaggerUIOptions } from './swagger/renderer'
+import { renderSwaggerUIOptions } from './swagger/renderer'
 
 type TestCase = [description: string, config: DistSwaggerUIOptions, expected: string]
 
@@ -189,29 +189,23 @@ describe('SwaggerUIOption Rendering', () => {
     ],
     [
       'filters out properties not in RENDER_TYPE_MAP',
-      { url: baseUrl, title: 'My API', customProperty: 'value' } as DistSwaggerUIOptions & { title: string },
-      `url: '${baseUrl}'`
+      { url: baseUrl, title: 'My API', customProperty: 'value' } as DistSwaggerUIOptions & {
+        title: string
+      },
+      `url: '${baseUrl}'`,
     ],
-    [
-      'filters out undefined values',
-      { url: baseUrl, layout: undefined },
-      `url: '${baseUrl}'`
-    ],
+    ['filters out undefined values', { url: baseUrl, layout: undefined }, `url: '${baseUrl}'`],
     [
       'handles multiple invalid properties gracefully',
-      { 
-        url: baseUrl, 
+      {
+        url: baseUrl,
         title: 'My API', // Not in RENDER_TYPE_MAP but used in HTML <title>
         presets: null as unknown, // Invalid type
-        withCredentials: true 
+        withCredentials: true,
       } as DistSwaggerUIOptions & { title: string },
-      `url: '${baseUrl}',withCredentials: true`
+      `url: '${baseUrl}',withCredentials: true`,
     ],
-    [
-      'handles empty input gracefully',
-      { url: '' },
-      `url: ''`
-    ]
+    ['handles empty input gracefully', { url: '' }, `url: ''`],
   ]
 
   it.each(commonTests)('renders correctly with %s', (_, input, expected) => {
