+++ packages/trpc-server/src/index.test.ts
@@ -1,7 +1,7 @@
 import { initTRPC } from '@trpc/server'
 import { Hono } from 'hono'
 import { z } from 'zod'
-import { trpcServer } from '../src'
+import { trpcServer } from '.'
 
 describe('tRPC Adapter Middleware', () => {
   const t = initTRPC.create()
