+++ packages/zod-openapi/src/handler.test-d.ts
@@ -1,8 +1,8 @@
 import type { MiddlewareHandler } from 'hono'
 import type { Equal, Expect } from 'hono/utils/types'
-import type { MiddlewareToHandlerType, OfHandlerType, RouteHandler } from '../src'
+import type { MiddlewareToHandlerType, OfHandlerType, RouteHandler } from '.'
 
-import { OpenAPIHono, createRoute, z } from '../src'
+import { OpenAPIHono, createRoute, z } from '.'
 
 describe('supports async handler', () => {
   const route = createRoute({
