+++ packages/zod-openapi/src/index.test-d.ts
@@ -2,8 +2,8 @@ import { createMiddleware } from 'hono/factory'
 import type { ExtractSchema } from 'hono/types'
 import type { Equal, Expect } from 'hono/utils/types'
 import { assertType, describe, it } from 'vitest'
-import { OpenAPIHono, createRoute, z } from '../src/index'
-import type { MiddlewareToHandlerType, OfHandlerType } from '../src/index'
+import { OpenAPIHono, createRoute, z } from './index'
+import type { MiddlewareToHandlerType, OfHandlerType } from './index'
 
 describe('Types', () => {
   const RequestSchema = z.object({
